backend/testing-requirements.txt
pytest>=8.0.0
pytest-asyncio>=0.23.5
httpx>=0.26.0
fastapi>=0.115.8
sqlalchemy>=2.0.38
alembic>=1.14.1
pydantic>=2.10.6
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
cryptography>=44.0.1
redis>=5.0.1
fakeredis>=2.20.1 